ParticipantList moved to separate folder
